Unique usernames, set session id on register, clean up extra files
