+++ app/controllers/users_controller.rb
@@ -11,8 +11,9 @@ class UsersController < ApplicationController
 
   def create
     @user = User.new(user_params)
-    
+
     if @user.save
+      session[:user_id] = @user.id
       redirect_to @user
     else
       render 'new'
