--- app/helpers/users_controller_helper.rb
@@ -1,2 +1,0 @@
-module UsersControllerHelper
-end
