+++ app/views/users/show.html.erb
@@ -1,1 +1,4 @@
 <h2>Hello <%= @user.username %>!</h2>
+
+<p><%= user_signed_in? %></p>
+
