+++ db/migrate/20160525160158_create_articles.rb
@@ -2,7 +2,7 @@ class CreateArticles < ActiveRecord::Migration
   def change
     create_table :articles do |t|
       t.string :title, null:false
-      t.integer :category_id
+      t.integer :category_id, null:false
       t.boolean :featured, default:false
 
       t.timestamps null: false
